Chschrae/remove nightfury refs (#62)
* Renamed files and scripts to properly reflect GGEC speaker

* Renamed files and scripts to properly reflect GGEC speaker

* updated readmes

* typo fix

* capitalized Windows and fixed typo
